Added filter in traits
