+++ Modules/Project/Entities/Project.php
@@ -2,6 +2,7 @@
 
 namespace Modules\Project\Entities;
 
+use App\Traits\Filters;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Modules\Client\Entities\Client;
@@ -12,7 +13,7 @@ use Modules\EffortTracking\Services\EffortTrackingService;
 
 class Project extends Model
 {
-    use HasFactory;
+    use HasFactory, Filters;
 
     protected $guarded = [];
 
@@ -63,30 +64,5 @@ class Project extends Model
         $monthlyEstimatedHours = $this->monthly_estimated_hours;
 
         return $monthlyEstimatedHours ? round($totalEffort / $monthlyEstimatedHours, 2) : 0;
-    }
-
-    public function scopeFilterByStatus($query, $status)
-    {
-        return $query->where('status', $status);
-    }
-
-    public function scopeFilterByName($query, $name)
-    {
-        return $query->where('name', 'LIKE', "%$name%");
-    }
-
-    public function scopeApplyFilter($query, array $filters)
-    {
-        foreach (array_filter($filters) as $type => $value) {
-            switch ($type) {
-                case 'status':
-                    $query->filterByStatus($value);
-                    break;
-                case 'name':
-                    $query->filterByName($value);
-            }
-        }
-
-        return $query;
     }
 }
